Refactor board to hexagonal architecture
