+++ src/main/kotlin/com/example/demo/domain/model/Post.kt
@@ -1,4 +1,4 @@
-package com.example.demo.model
+package com.example.demo.domain.model
 
 import java.util.UUID
